refactor: make net ectually track bees instead of a number

+++ src/GameObjects/BeeKeeper/states/BeeKeeperChaseState.hpp
@@ -16,7 +16,7 @@ public:
 
 private:
 
-    const int _max_bees_in_net = 10;
+    const size_t _max_bees_in_net = 10;
 
     void _recalculate_path();
     Vertex* _get_target_vertex();
